refactor(app): split dashboard scenario tests
